Add traverse method to SourceRootTrie
This will return a set of named tuples of all valid paths (with filesystem
globbing included), and the languages/categories each path applies to, that a
given SourceRootTrie instance describes. This will be useful for
implementing a v2 engine-compatible way of getting a list of all source
files, but is a standalone change in and of itself.